rename queue/uiReducer files
