+++ client/app/queue/QueueListView.jsx
@@ -8,7 +8,7 @@ import StatusMessage from '../components/StatusMessage';
 import QueueTable from './QueueTable';
 import AppSegment from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/AppSegment';
 
-import { resetErrorMessages } from './uiReducer/actions';
+import { resetErrorMessages } from './uiReducer/uiActions';
 import { clearCaseSelectSearch } from '../reader/CaseSelect/CaseSelectActions';
 
 import { fullWidth } from './constants';
